+++ PRUNner/Backend/Enums/ExchangePriceType.cs
@@ -7,7 +7,7 @@ namespace PRUNner.Backend.Enums
         Bid, 
         MMBuy,
         MMSell,
-        Worse,
-        Best,
+        WorstCase,
+        BestCase,
     }
 }